% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsyc.R
\docType{package}
\name{apaTables}
\alias{apaTables}
\alias{ggpsyc}
\alias{apaTables-package}
\title{Greate ITNS graphs}
\description{
Users of the Introduction to the New Statistics textbook may wish to create graphs similar to those in the book - which this package faciliates.
}
\details{
Bugs and feature requests can be reported at: \url{https://github.com/dstanley4/ggpsyc/issues}

\tabular{ll}{
Package: \tab ggpsyc\cr
Type: \tab Package\cr
Version: \tab 0.0.9\cr
Date: \tab 2019-06-15\cr
}
}
\author{
\tabular{ll}{
Author: \tab David J. Stanley \email{dstanley@uoguelph.ca}\cr
Maintainer: \tab David J. Stanley \email{dstanley@uoguelph.ca}
}
}
