% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_combine.R
\name{geom_two_group_between}
\alias{geom_two_group_between}
\title{Create a plot for independent groups t-test as per Introduction to the New Statistics}
\usage{
geom_two_group_between()
}
\description{
Create a plot for independent groups t-test as per Introduction to the New Statistics
}
