% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mean_diff_repeated.R
\name{stat_mean_diff_repeated}
\alias{stat_mean_diff_repeated}
\title{This is the documentation for stat_catseye. remember to account for missing data}
\usage{
stat_mean_diff_repeated(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, group1 = NA, group2 = NA,
  alternative = "two.sided", text_size = 3, ...)
}
\arguments{
\item{mapping}{asdfasd}

\item{data}{adfasd}

\item{geom}{asdfasd}

\item{position}{asdfasd}

\item{na.rm}{adsfasdf}

\item{show.legend}{asdfads}

\item{inherit.aes}{adfasd}

\item{group1}{asdsf}

\item{group2}{asdsf}

\item{...}{adfasdf}

\item{width}{asdf}
}
\description{
This is the documentation for stat_catseye. remember to account for missing data
}
