% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_definitions.R
\name{stat_two_group_ci}
\alias{stat_two_group_ci}
\title{This is the documentation for stat_catseye. remember to account for missing data}
\usage{
stat_two_group_ci(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, paired = FALSE, var.equal = TRUE,
  text_size = 3, level = 0.95, ...)
}
\arguments{
\item{mapping}{asdfasd}

\item{data}{adfasd}

\item{geom}{asdfasd}

\item{position}{asdfasd}

\item{na.rm}{adsfasdf}

\item{show.legend}{asdfads}

\item{inherit.aes}{adfasd}

\item{...}{adfasdf}

\item{group1}{asdsf}

\item{group2}{asdsf}

\item{width}{asdf}
}
\description{
This is the documentation for stat_catseye. remember to account for missing data
}
